Move sound scripting classes into separate file
